+++ cmd/forklift/cache.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"fmt"
 	"sort"
 	"strings"
@@ -10,6 +11,7 @@ import (
 
 	"github.com/PlanktoScope/forklift/internal/app/forklift"
 	"github.com/PlanktoScope/forklift/internal/app/forklift/workspace"
+	"github.com/PlanktoScope/forklift/internal/clients/docker"
 )
 
 // ls-repo
@@ -182,6 +184,77 @@ func cacheShowPkgAction(c *cli.Context) error {
 	}
 	printCachedPkg(pkg)
 	return nil
+}
+
+// ls-img
+
+func cacheLsImgAction(c *cli.Context) error {
+	client, err := docker.NewClient()
+	if err != nil {
+		return errors.Wrap(err, "couldn't make Docker API client")
+	}
+
+	imgs, err := client.ListImages(context.Background())
+	if err != nil {
+		return errors.Wrapf(err, "couldn't list local Docker images")
+	}
+	sort.Slice(imgs, func(i, j int) bool {
+		return imgs[i].Repository < imgs[j].Repository
+	})
+	for _, img := range imgs {
+		fmt.Printf("%s: %s\n", img.ID, img.Repository)
+	}
+	return nil
+}
+
+// show-img
+
+func cacheShowImgAction(c *cli.Context) error {
+	client, err := docker.NewClient()
+	if err != nil {
+		return errors.Wrap(err, "couldn't make Docker API client")
+	}
+
+	imageHash := c.Args().First()
+	image, err := client.InspectImage(context.Background(), imageHash)
+	if err != nil {
+		return errors.Wrapf(err, "couldn't inspect image %s", imageHash)
+	}
+	printImg(image)
+	return nil
+}
+
+func printImg(img docker.Image) {
+	fmt.Printf("Docker container image: %s\n", img.ID)
+	fmt.Print("  Provided by container image repository: ")
+	if img.Repository == "" {
+		fmt.Print("(none)")
+	} else {
+		fmt.Print(img.Repository)
+	}
+	fmt.Println()
+
+	fmt.Print("    Repo tags:")
+	if len(img.Inspect.RepoTags) == 0 {
+		fmt.Print(" (none)")
+	}
+	fmt.Println()
+	for _, tag := range img.Inspect.RepoTags {
+		fmt.Printf("      %s\n", tag)
+	}
+
+	fmt.Print("    Repo digests:")
+	if len(img.Inspect.RepoDigests) == 0 {
+		fmt.Print(" (none)")
+	}
+	fmt.Println()
+	for _, digest := range img.Inspect.RepoDigests {
+		fmt.Printf("      %s\n", digest)
+	}
+
+	fmt.Printf("  Created: %s\n", img.Inspect.Created)
+	const mbConversion = 1024 * 1024
+	fmt.Printf("  Size: %.1f MiB\n", float32(img.Inspect.Size)/mbConversion)
 }
 
 // rm